==> SMS-SPAM/3 - Extract, Review _ Combine Datasets.ipynb <==
import pathlib

BASE_DIR = pathlib.Path().resolve().parent #Check: /Users/kristoffervarslott/Documents/Python.py/AI_API
SMS_SPAM_DIR = pathlib.Path().resolve()    #Check: /Users/kristoffervarslott/Documents/Python.py/AI_API/SMS-SPAM

DATASETS_DIR = BASE_DIR / 'Datasets' #Check: /Users/kristoffervarslott/Documents/Python.py/AI_API/Datasets
ZIPS_DIR = DATASETS_DIR / 'Zips'     #Check: /Users/kristoffervarslott/Documents/Python.py/AI_API/Datasets/Zips

ZIPS_DIR.mkdir(exist_ok = True, parents = True)

SPAM_SMS_ZIP_PATH = ZIPS_DIR / 'sms-spam-dataset.zip'
SPAM_YOUTUBE_ZIP_PATH = ZIPS_DIR / 'youtube-spam-dataset.zip'

SMS_SPAM_ZIP = 'https://archive.ics.uci.edu/ml/machine-learning-databases/00228/smsspamcollection.zip'
YOUTUBE_SPAM_ZIP = 'https://archive.ics.uci.edu/ml/machine-learning-databases/00380/YouTube-Spam-Collection-v1.zip'

!curl $SMS_SPAM_ZIP -o $SPAM_SMS_ZIP_PATH
!curl $YOUTUBE_SPAM_ZIP -o $SPAM_YOUTUBE_ZIP_PATH


SPAM_CLASSIFIER_DIR = DATASETS_DIR / 'Spam-Classifier'

SMS_SPAM_DIR = SPAM_CLASSIFIER_DIR / 'Sms-Spam'
SMS_SPAM_DIR.mkdir(exist_ok = True, parents = True)

YOUTUBE_SPAM_DIR = SPAM_CLASSIFIER_DIR / 'Youtube-Spam'
YOUTUBE_SPAM_DIR.mkdir(exist_ok = True, parents = True)

EXPORT_DIR = DATASETS_DIR / 'Exports'
EXPORT_DIR.mkdir(exist_ok = True, parents = True)
SPAM_DATASETS_DIR = EXPORT_DIR / 'Spam_Dataset.csv'

!unzip -o $SPAM_SMS_ZIP_PATH -d $SMS_SPAM_DIR
!unzip -o $SPAM_YOUTUBE_ZIP_PATH -d $YOUTUBE_SPAM_DIR

import pandas as pd

#All files contained in SMS_SPAM_DIR: 
sms_spam_input_path = SMS_SPAM_DIR / 'SMSSpamCollection'

sms_df = pd.read_csv(sms_spam_input_path, sep='\t', header=None) #Setting header None in order to include all datapoints
sms_df.columns  = ['LABEL', 'CONTENT']
sms_df['SOURCE'] = 'sms-spam'
sms_df['CLASS'] = sms_df['LABEL'].apply(lambda x: '1' if str(x) == 'spam' else '0')
sms_df = sms_df[["CONTENT", "CLASS",'LABEL', "SOURCE"]]
sms_df

#All files contained in YOUTUBE_SPAM_DIR: 
youtube_spam_df = []

#Looping through YOUTUBE_SPAM_DIR containing a number of .csv files - furthermore, appending these to a list. 
for path in YOUTUBE_SPAM_DIR.glob('*.csv'): 

    df_raw = pd.read_csv(path)
    
    df_raw_copy = df_raw.copy()
    df_raw_copy['SOURCE'] = 'Spam-youtube'
    df_raw_copy = df_raw_copy.drop(columns=['COMMENT_ID', 'AUTHOR', 'DATE']) #Droppped irrelevant data from database
    df_raw_copy['LABEL'] = df_raw_copy['CLASS'].apply(lambda x: 'spam' if str(x) == '1' else 'ham')
    df_raw_copy = df_raw_copy[["CONTENT", "CLASS",'LABEL', "SOURCE"]]
    youtube_spam_df.append(df_raw_copy)

youtube_df =  pd.concat(youtube_spam_df, ignore_index=True) #concat, merging a list of pd.read_csv into one
youtube_df

df_final =  pd.concat([sms_df,youtube_df])
df_final

df_final.to_csv(SPAM_DATASETS_DIR, index=False) #Pushing Combined Dataset to .csv at SPAM_DATASETS_DIR for future work.
